feat: add flags to macro, add axes titles
